factory/execution/requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
replicate>=0.25.0
numpy>=1.24.0
scipy>=1.11.0
pydub>=0.25.0
